Fix weird texture mapping issue

+++ linuxdoom-1.10/p_pspr.c
@@ -526,7 +526,7 @@ A_Saw
 			     linetarget->x, linetarget->y);
     if (angle - player->mo->angle > ANG180)
     {
-	if ((int)angle - (int)player->mo->angle < -ANG90/20)
+	if ((int)(angle - player->mo->angle) < -ANG90/20)
 	    player->mo->angle = angle + ANG90/21;
 	else
 	    player->mo->angle -= ANG90/20;
