+++ linuxdoom-1.10/r_plane.c
@@ -131,7 +131,7 @@ R_MapPlane
     if (x2 <  x1
 	||  x1 <  0
 	||  x2 >= viewwidth
-	|| (int)(unsigned)y > viewheight)
+	||  y  >  viewheight)
     {
 	    I_Error ("R_MapPlane: %i, %i at %i", x1, x2, y);
     }
